Split wave executor modules and dedupe run advancement lifecycle
