+++ rust/loopflow/src/lfd/executor/wave/fork.rs
@@ -1,0 +1,280 @@
+use std::path::Path;
+
+use anyhow::{anyhow, Result};
+use tokio::sync::mpsc;
+use tokio_util::sync::CancellationToken;
+use tracing::{debug, error, info, warn};
+
+use crate::engine::agent::build_agent_command;
+use crate::engine::flow::{ConcreteFork, ConcreteItem, ConcreteStep};
+use crate::engine::worktree::{create_worktree, remove_worktree};
+use crate::lfd::config::ExecutorType;
+use crate::lfd::id::LfdId;
+use crate::lfd::store::{ForkRun, ForkRunStatus};
+use crate::lfd::types::{Wave, WaveRun};
+
+use super::launch::AgentLaunchRequest;
+use super::WaveExecutor;
+use crate::lfd::executor::helpers::{build_step_prompt, fork_worktree_path};
+
+impl WaveExecutor {
+    pub(super) async fn run_choose(
+        &self,
+        wave: &Wave,
+        run: &mut WaveRun,
+        plan: &[ConcreteItem],
+        fork: &ConcreteFork,
+    ) -> Result<()> {
+        if fork.branches.is_empty() {
+            self.fail_run(run, wave, "fork has no branches".to_string())?;
+            return Ok(());
+        }
+
+        let selected = fork
+            .branches
+            .first()
+            .ok_or_else(|| anyhow!("fork has no branches"))?
+            .clone();
+
+        if selected.step.interactive.unwrap_or(false) {
+            self.fail_run(
+                run,
+                wave,
+                "interactive fork branches are not supported".to_string(),
+            )?;
+            return Ok(());
+        }
+
+        let exit_code = self.run_step(wave, run, &selected).await?;
+        if exit_code != 0 {
+            self.fail_run(
+                run,
+                wave,
+                format!("fork step {} failed", selected.step.name),
+            )?;
+            return Ok(());
+        }
+
+        self.advance_run_step(run, plan, &wave.id)?;
+        Ok(())
+    }
+
+    pub(super) async fn run_fork(
+        &self,
+        wave: &Wave,
+        run: &mut WaveRun,
+        plan: &[ConcreteItem],
+        fork: &ConcreteFork,
+    ) -> Result<()> {
+        if self.executor_type == ExecutorType::Docker {
+            self.fail_run(
+                run,
+                wave,
+                "fork(select=all) is not supported by the docker executor yet".to_string(),
+            )?;
+            return Ok(());
+        }
+
+        let mut fork_runs = Vec::new();
+        for (index, branch) in fork.branches.iter().enumerate() {
+            if branch.step.interactive.unwrap_or(false) {
+                self.fail_run(
+                    run,
+                    wave,
+                    "interactive fork branches are not supported".to_string(),
+                )?;
+                return Ok(());
+            }
+
+            let fork_worktree = fork_worktree_path(run, index as u32);
+            if !Path::new(&fork_worktree).exists() {
+                debug!(
+                    run_id = %run.id,
+                    branch_index = index,
+                    step = %branch.step.name,
+                    worktree = %fork_worktree,
+                    "creating fork worktree"
+                );
+                create_worktree(
+                    Path::new(&run.snapshot.repo),
+                    Path::new(&fork_worktree),
+                    &format!("{}-fork-{}", run.id, index),
+                )?;
+            }
+
+            let fork_run = ForkRun {
+                id: LfdId::new(),
+                wave_run_id: run.id.clone(),
+                step_index: run.step_index,
+                branch_index: index as u32,
+                status: ForkRunStatus::Pending,
+                worktree: fork_worktree,
+            };
+            self.store.upsert_fork_run(&fork_run)?;
+            fork_runs.push((fork_run, branch.clone()));
+        }
+
+        let cancel = CancellationToken::new();
+        let (tx, mut rx) = mpsc::channel(fork_runs.len());
+        let mut handles = Vec::new();
+
+        let wave_directions = run.snapshot.direction.clone();
+        for (fork_run, step) in fork_runs.iter() {
+            let executor = self.clone();
+            let scheduler = self.scheduler.clone();
+            let cancel = cancel.clone();
+            let tx = tx.clone();
+            let fork_wave_id = wave.id.clone();
+            let wave_run_id = run.id.clone();
+            let wave_repo = run.snapshot.repo.clone();
+            let worktree = fork_run.worktree.clone();
+            let fork_run_id = fork_run.id.clone();
+            let fork_run = fork_run.clone();
+            let step = step.clone();
+            let wave_directions = wave_directions.clone();
+
+            let handle = tokio::spawn(async move {
+                if cancel.is_cancelled() {
+                    return;
+                }
+
+                loop {
+                    let (acquired, _) = scheduler.acquire(fork_run_id.as_str()).await;
+                    if acquired {
+                        break;
+                    }
+                    tokio::time::sleep(std::time::Duration::from_millis(200)).await;
+                }
+
+                let _ = executor.store.upsert_fork_run(&ForkRun {
+                    status: ForkRunStatus::Running,
+                    ..fork_run.clone()
+                });
+
+                debug!(
+                    fork_run_id = %fork_run_id,
+                    step = %step.step.name,
+                    worktree = %worktree,
+                    directions = ?wave_directions,
+                    "building fork branch prompt"
+                );
+                let prompt = build_step_prompt(
+                    &worktree,
+                    &step,
+                    &wave_directions,
+                    None,
+                    Some((&executor.store, &fork_wave_id)),
+                    None,
+                );
+                let (prompt, model, launch) = match prompt {
+                    Ok(result) => result,
+                    Err(err) => {
+                        error!(
+                            fork_run_id = %fork_run_id,
+                            step = %step.step.name,
+                            error = %err,
+                            "fork branch prompt build failed"
+                        );
+                        let _ = tx.send((fork_run_id.to_string(), Err(err))).await;
+                        scheduler.release(fork_run_id.as_str());
+                        return;
+                    }
+                };
+                let cmd = build_agent_command(&model, &prompt, &launch);
+                info!(
+                    fork_run_id = %fork_run_id,
+                    step = %step.step.name,
+                    model = %model,
+                    cmd_len = cmd.len(),
+                    "launching fork branch agent"
+                );
+
+                let result = executor
+                    .launch_agent(AgentLaunchRequest {
+                        wave_id: fork_wave_id.clone(),
+                        wave_run_id: wave_run_id.clone(),
+                        repo: wave_repo.clone(),
+                        worktree: worktree.clone(),
+                        step: step.clone(),
+                        model,
+                        cmd,
+                    })
+                    .await
+                    .map(|outcome| outcome.exit_code);
+
+                let status = match &result {
+                    Ok(0) => {
+                        info!(fork_run_id = %fork_run_id, step = %step.step.name, "fork branch completed");
+                        ForkRunStatus::Completed
+                    }
+                    Ok(code) => {
+                        warn!(fork_run_id = %fork_run_id, step = %step.step.name, exit_code = code, "fork branch failed");
+                        ForkRunStatus::Failed
+                    }
+                    Err(err) => {
+                        error!(fork_run_id = %fork_run_id, step = %step.step.name, error = %err, "fork branch error");
+                        ForkRunStatus::Failed
+                    }
+                };
+                let _ = executor.store.upsert_fork_run(&ForkRun {
+                    status,
+                    ..fork_run.clone()
+                });
+                let _ = tx.send((fork_run_id.to_string(), result)).await;
+                scheduler.release(fork_run_id.as_str());
+            });
+
+            handles.push(handle);
+        }
+
+        let mut failures = None;
+        let mut completed = 0usize;
+        let total = fork_runs.len();
+        debug!(run_id = %run.id, total_branches = total, "waiting for fork results");
+        while let Some((fork_id, result)) = rx.recv().await {
+            match result {
+                Ok(0) => {
+                    completed += 1;
+                    debug!(run_id = %run.id, completed, total, "fork branch done");
+                    if completed == total {
+                        break;
+                    }
+                }
+                Ok(code) => {
+                    failures = Some(format!("fork branch {} exited with code {}", fork_id, code));
+                    break;
+                }
+                Err(err) => {
+                    failures = Some(format!("fork branch {} error: {}", fork_id, err));
+                    break;
+                }
+            }
+        }
+
+        if let Some(error) = failures {
+            error!(run_id = %run.id, error = %error, "fork failed");
+            cancel.cancel();
+            for handle in handles {
+                handle.abort();
+            }
+            self.cleanup_fork(run, &fork_runs).await;
+            self.fail_run(run, wave, error)?;
+            return Ok(());
+        }
+
+        self.cleanup_fork(run, &fork_runs).await;
+        self.advance_run_step(run, plan, &wave.id)?;
+        Ok(())
+    }
+
+    async fn cleanup_fork(&self, run: &WaveRun, fork_runs: &[(ForkRun, ConcreteStep)]) {
+        for (fork_run, _) in fork_runs {
+            let worktree_path = Path::new(&fork_run.worktree);
+            if worktree_path.join(".git").exists() {
+                let _ = remove_worktree(worktree_path, true);
+            }
+            self.scheduler.release(fork_run.id.as_str());
+        }
+        let _ = self.store.delete_fork_runs(&run.id, run.step_index);
+    }
+}
